requirements.txt
numpy~=1.21.4
opencv-python~=4.5.4.58
imutils~=0.5.4
matplotlib~=3.4.3
Pillow~=8.4.0
tqdm~=4.62.3
pandas~=1.3.4
seaborn~=0.11.2
predictr~=0.1.25
scipy~=1.7.3
plotly~=5.3.1
dash~=2.0.0
moviepy~=1.0.3
celluloid~=0.2.0
ipython~=8.2.0